append plugin autoPrefix to default prefix
Using `prefix` in the default autoload options applies a prefix for all plugins loaded by this module. This makes it so any modules then loaded by autoload appends the modules autoPrefix to the end of the default autoload prefix.

It also adds a prefixOverride to override the defaultPrefix (and autoPrefix).

Tests and docs to follow.